- finalized ExamController and ExamService.

+++ common/database/src/main/java/com/sebrown2023/repository/ExamRepository.java
@@ -9,5 +9,5 @@ import java.util.Set;
 @Repository
 public interface ExamRepository extends JpaRepository<Exam, Long> {
     Exam findExamById(long id);
-    Set<Exam> findExamsByExaminerIdEquals(int examinerId);
+    Set<Exam> findExamsByExaminerIdEquals(long examinerId);
 }
